64 bit warnings-- - need to install a 64bit linux after holiday here... :)
git-svn-id: file:///.data/svn_convert/svn@6865 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-QString/sim/cmddef.cpp
@@ -123,7 +123,7 @@ void *CommandsDefPrivate::processEvent(Event *e)
         }
         break;
     case EventCommandRemove:
-        if (delCommand((unsigned long)(e->param())))
+        if (delCommand((unsigned)(e->param())))
             cfg.clear();
         break;
     }
